Move CompactionWatcher and reuse in Compactor
